reducer 코드 분리
