+++ src/app.js
@@ -4,6 +4,7 @@ import styled from "styled-components";
 import MyNewsTitle from "./MyNewsTitle.jsx";
 import ListUi from "./ListUi.jsx";
 import CardUi from "./CardUi.jsx";
+import {initState, pageReducer} from './pageReducer';
 
 const ContentsWrap = styled.div`
   width: 970px;
@@ -11,39 +12,6 @@ const ContentsWrap = styled.div`
 `;
 
 const DataContext = React.createContext({});
-
-const numPerListPage = 6;
-const numPerCardPage = 18;
-const initState = {
-  limit: numPerListPage,
-  offset: 0,
-  type: 'LIST'
-};
-const pageReducer = (state, {type}) => {
-  const state_copy = {...state};
-  switch (type) {
-    case 'List':
-      state_copy.type = 'LIST';
-      state_copy.limit = numPerListPage;
-      state_copy.offset = 0;
-      return state_copy;
-    case 'Card':
-      state_copy.type = 'CARD';
-      state_copy.limit = numPerCardPage;
-      state_copy.offset = 0;
-      return state_copy;
-    case 'goToNext':
-      state_copy.offset = state_copy.offset + 1;
-      return state_copy;
-    case 'goToPrev':
-      state_copy.offset = state_copy.offset === 0 ? 0 : state_copy.offset - 1;
-      return state_copy;
-    case 'init':
-      return initState;
-    default:
-      break;
-  }
-};
 
 function App() {
   const [newsData, setNewsData] = useState([]);
@@ -57,8 +25,7 @@ function App() {
 
   useEffect(() => {
     getNews().then(news => {
-      const offset = state.offset;
-      const limit = state.limit;
+      const {offset, limit} = {...state};
       const paging = offset <= limit ? offset : parseInt(offset / limit);
       const data = news.slice(paging * limit, (paging + 1) * limit);
       if(data.length === 0) return;
